UPSTREAM: 11475: fix(backend): modelToCRDTrigger was not including periodic schedule correctly (#11475)
---------

(cherry picked from commit 97acacbd2a0b72d442398ca04382ac1e6d9aa37f)
